Unit tests added

+++ src/main/java/com/qa/account/service/TaskService.java
@@ -1,9 +1,10 @@
 package com.qa.account.service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.qa.account.dto.TaskDTO;
@@ -15,35 +16,25 @@ import com.qa.account.persistence.repo.TaskRepo;
 public class TaskService {
 
 	private TaskRepo repo;
-	private ModelMapper mapper;
+	private Mapper<Task, TaskDTO> mapper;
 
+	@Autowired
 	public TaskService(TaskRepo repo, ModelMapper mapper) {
 		super();
 		this.repo = repo;
-		this.mapper = mapper;
-	}
-	
-	private TaskDTO mapToDTO(Task task) {
-		return this.mapper.map(task, TaskDTO.class);
+		this.mapper = (Task task) -> mapper.map(task, TaskDTO.class);
 	}
 
 	public TaskDTO create(Task task) {
-//		return this.repo.save(task);
-		Task saved = this.repo.save(task);
-		return this.mapToDTO(saved);
+		return this.mapper.mapToDTO(this.repo.save(task));
 	}
 
 	public List<TaskDTO> read() {
-		List<TaskDTO> dtos = new ArrayList<>();
-		for (Task task : this.repo.findAll()) {
-			dtos.add(this.mapToDTO(task));
-		}
-		return dtos;
-//		return this.repo.findAll().stream().map(this::mapToDTO).collect(Collectors.toList()); // same as above
+		return this.repo.findAll().stream().map(this.mapper::mapToDTO).collect(Collectors.toList());
 	}
 
-	public Task read(long taskId) {
-		return this.repo.findById(taskId).orElseThrow(() -> new TaskNotFoundException());
+	public TaskDTO readId(Long taskId) {
+		return this.mapper.mapToDTO(this.repo.findById(taskId).orElseThrow(TaskNotFoundException::new));
 	}
 
 	public Task update(Task task, Long taskId) {
@@ -58,6 +49,9 @@ public class TaskService {
 	}
 
 	public boolean delete(Long taskId) {
+		if (!this.repo.existsById(taskId)) {
+			throw new TaskNotFoundException();
+		}
 		this.repo.deleteById(taskId);
 		return this.repo.existsById(taskId);
 	}
